requirements.txt
pytelegrambotapi==4.9.0
numpy==1.24.3
faster-whisper==0.7.0
ffmpeg-python==0.2.0